+++ CHANGELOG.md
@@ -12,6 +12,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 -->
 
 ## [Unreleased]
+### Changed
+- Support FreeBSD by manually setting `PAM_SILENT`
 
 ## [1.0.0-alpha3] - 2021-03-28
 ### Changed